requirements.txt
Pillow==9.4.0

ttkthemes~=3.2.2